% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient-package.R
\docType{package}
\name{ambient-package}
\alias{ambient}
\alias{ambient-package}
\title{ambient: A Generator of Multidimensional Noise}
\description{
Generation of natural looking noise has many application within 
    simulation, procedural generation, and art, to name a few. The 'ambient' 
    package provides an interface to the 'FastNoise' C++ library and allows for
    efficient generation of perlin, simplex, worley, cubic, value, and white 
    noise with optional pertubation in either 2, 3, or 4 (in case of simplex and
    white noise) dimensions.
}
\references{
\url{https://github.com/Auburns/FastNoise}
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

Authors:
\itemize{
  \item Jordan Peck (Developer of FastNoise)
}

}
