% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{trans_affine}
\alias{trans_affine}
\alias{rotate}
\alias{stretch}
\alias{shear}
\alias{translate}
\alias{reflect}
\title{Apply linear transformation to a long_grid}
\usage{
trans_affine(grid, ..., x, y, output = NULL)

rotate(angle = 0)

stretch(x0 = 0, y0 = 0)

shear(x0 = 0, y0 = 0)

translate(x0 = 0, y0 = 0)

reflect(x0 = 0, y0 = 0)
}
\arguments{
\item{grid}{A long_grid object}

\item{...}{A sequence of transformations}

\item{x, y}{The unquoted columns holding the x and y coordinates to transform}

\item{output}{The column names to save the transformed coordinates in. If
\code{NULL} the names will be derived from \code{x} and \code{y}.}

\item{angle}{An angle in radians}

\item{x0}{the transformation magnitude in the x-direction}

\item{y0}{the transformation magnitude in the x-direction}
}
\description{
This function allows you to calculate linear transformations of coordinates
in a long_grid object. You can either pass in a transformation matrix or a
trans object as produced by \code{ggforce::linear_trans(...)}. The latter makes it
easy to stack multiple transformations into one, but require the ggforce
package.
}
\section{Linear Transformations}{

The following transformation matrix constructors are supplied, but you can
also provide your own 3x3 matrices to \code{translate()}
\itemize{
\item \code{rotate()}: Rotate coordinates by \code{angle} (in radians) around the center
counter-clockwise.
\item \code{stretch()}: Stretches the x and/or y dimension by multiplying it with
\code{x0}/\code{y0}.
\item \code{shear()}: Shears the x and/or y dimension by \code{x0}/\code{y0}.
\item \code{translate()}: Moves coordinates by \code{x0}/\code{y0}.
\item \code{reflect()}: Reflects coordinates through the line that goes through \code{0, 0}
and \code{x0, y0}.
}
}

\examples{
grid <- long_grid(seq(1, 10, length.out = 1000), seq(1, 10, length.out = 1000))
grid <- trans_affine(grid, rotate(pi/3), shear(-2), rotate(-pi/3), x = x, y = y)
grid$chess <- gen_checkerboard(grid$x, grid$y)

plot(as.raster(grid, chess))

}
