% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient-package.R
\docType{package}
\name{ambient-package}
\alias{ambient}
\alias{ambient-package}
\title{ambient: A Generator of Multidimensional Noise}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Generation of natural looking noise has many application within 
    simulation, procedural generation, and art, to name a few. The 'ambient' 
    package provides an interface to the 'FastNoise' C++ library and allows for
    efficient generation of perlin, simplex, worley, cubic, value, and white 
    noise with optional pertubation in either 2, 3, or 4 (in case of simplex and
    white noise) dimensions.
}
\references{
\url{https://github.com/Auburns/FastNoise}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ambient.data-imaginist.com}
  \item \url{https://github.com/thomasp85/ambient}
  \item Report bugs at \url{https://github.com/thomasp85/ambient/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Jordan Peck (Developer of FastNoise)
}

}
